chore: misc cleanup

+++ packages/db-postgres/src/deleteMany.ts
@@ -7,7 +7,6 @@ import toSnakeCase from 'to-snake-case'
 import type { PostgresAdapter } from './types'
 
 import { findMany } from './find/findMany'
-import { transform } from './transform/read'
 
 export const deleteMany: DeleteMany = async function deleteMany(
   this: PostgresAdapter,
@@ -17,7 +16,7 @@ export const deleteMany: DeleteMany = async function deleteMany(
   const collectionConfig = this.payload.collections[collection].config
   const tableName = toSnakeCase(collection)
 
-  const { docs } = await findMany({
+  const result = await findMany({
     adapter: this,
     fields: collectionConfig.fields,
     limit: 0,
@@ -31,17 +30,13 @@ export const deleteMany: DeleteMany = async function deleteMany(
 
   const ids = []
 
-  const result = docs.map((data) => {
+  result.docs.forEach((data) => {
     ids.push(data.id)
-
-    return transform({
-      config: this.payload.config,
-      data,
-      fields: collectionConfig.fields,
-    })
   })
 
-  await db.delete(this.tables[tableName]).where(inArray(this.tables[tableName].id, ids))
+  if (ids.length > 0) {
+    await db.delete(this.tables[tableName]).where(inArray(this.tables[tableName].id, ids))
+  }
 
   return result
 }
