+++ packages/db-postgres/src/deleteVersions.ts
@@ -8,7 +8,6 @@ import toSnakeCase from 'to-snake-case'
 import type { PostgresAdapter } from './types'
 
 import { findMany } from './find/findMany'
-import { transform } from './transform/read'
 
 export const deleteVersions: DeleteVersions = async function deleteVersion(
   this: PostgresAdapter,
@@ -34,17 +33,13 @@ export const deleteVersions: DeleteVersions = async function deleteVersion(
 
   const ids = []
 
-  const result = docs.map((data) => {
-    ids.push(data.id)
-
-    return transform({
-      config: this.payload.config,
-      data,
-      fields: collectionConfig.fields,
-    })
+  docs.forEach((doc) => {
+    ids.push(doc.id)
   })
 
-  await db.delete(this.tables[tableName]).where(inArray(this.tables[tableName].id, ids))
+  if (ids.length > 0) {
+    await db.delete(this.tables[tableName]).where(inArray(this.tables[tableName].id, ids))
+  }
 
-  return result
+  return docs
 }
